chore(Entity): Only show status indicator for devices
